// src/api-client/models/budget-category-relation-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial Aide Backend
 * API Backend for Financial Aide open-source budgeting system
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BudgetCategoryResponse } from './budget-category-response';
import { BudgetResponse } from './budget-response';
/**
 * 
 * @export
 * @interface BudgetCategoryRelationResponse
 */
export interface BudgetCategoryRelationResponse {
    /**
     * 
     * @type {BudgetResponse}
     * @memberof BudgetCategoryRelationResponse
     */
    budget: BudgetResponse;
    /**
     * 
     * @type {BudgetCategoryResponse}
     * @memberof BudgetCategoryRelationResponse
     */
    category: BudgetCategoryResponse;
    /**
     * 
     * @type {string}
     * @memberof BudgetCategoryRelationResponse
     */
    amount: string;
    /**
     * 
     * @type {boolean}
     * @memberof BudgetCategoryRelationResponse
     */
    is_percentage: boolean;
    /**
     * 
     * @type {number}
     * @memberof BudgetCategoryRelationResponse
     */
    id: number;
}