// src/api-client/models/interval-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial Aide Backend
 * API Backend for Financial Aide open-source budgeting system
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * * `yearly` - Yearly * `monthly` - Monthly * `weekly` - Weekly
 * @export
 * @enum {string}
 */
export enum IntervalEnum {
    Yearly = 'yearly',
    Monthly = 'monthly',
    Weekly = 'weekly'
}